Add IPv6 support if enabled in Docker
This adds IPv6 support if enabled in Docker configuration.
Subnets sized as /64 are created in similar manner as IPv4.

Note that IPv6 requires a GUA and either NATv6 or valid routing
to be able to navigate outside Docker, as Docker doesn't configure NAT
as it does with IPv4.
Alternatively, ULA can be used locally, if no external connectivity
is required.